+++ packages/backend-data/src/factory.ts
@@ -26,7 +26,12 @@ import {
   isUsingDefaultApiKeyAuth,
 } from './convert_authorization_modes.js';
 import { validateAuthorizationModes } from './validate_authorization_modes.js';
-import { AmplifyUserError, CDKContextKey } from '@aws-amplify/platform-core';
+import {
+  AmplifyError,
+  AmplifyFault,
+  AmplifyUserError,
+  CDKContextKey,
+} from '@aws-amplify/platform-core';
 import { Aspects, IAspect } from 'aws-cdk-lib';
 import { convertJsResolverDefinition } from './convert_js_resolvers.js';
 import { AppSyncPolicyGenerator } from './app_sync_policy_generator.js';
@@ -133,24 +138,16 @@ class DataGenerator implements ConstructContainerEntryGenerator {
     }
 
     let authorizationModes;
-
-    /**
-     * TODO - remove this after the data construct does work to remove the need for allow-listed IAM roles
-     */
-    const functionSchemaAccessRoles = functionSchemaAccess.map(
-      (accessEntry) =>
-        accessEntry.resourceProvider.getInstance(this.getInstanceProps)
-          .resources.lambda.role!
-    );
-
     try {
       authorizationModes = convertAuthorizationModesToCDK(
         this.getInstanceProps,
         this.providedAuthConfig,
-        this.props.authorizationModes,
-        functionSchemaAccessRoles
+        this.props.authorizationModes
       );
     } catch (error) {
+      if (error instanceof AmplifyError) {
+        throw error;
+      }
       throw new AmplifyUserError<AmplifyDataError>(
         'InvalidSchemaAuthError',
         {
@@ -194,23 +191,33 @@ class DataGenerator implements ConstructContainerEntryGenerator {
       ...propsFunctions,
       ...lambdaFunctions,
     });
-    const amplifyApi = new AmplifyData(scope, this.defaultName, {
-      apiName: this.props.name,
-      definition: amplifyGraphqlDefinition,
-      authorizationModes,
-      outputStorageStrategy: this.outputStorageStrategy,
-      functionNameMap,
-      translationBehavior: {
-        sandboxModeEnabled,
-        /**
-         * The destructive updates should be always allowed in backend definition and not to be controlled on the IaC
-         * The CI/CD check should take the responsibility to validate if any tables are being replaced and determine whether to execute the changeset
-         */
-        allowDestructiveGraphqlSchemaUpdates: true,
-      },
-    });
+    let amplifyApi = undefined;
 
-    /**
+    try {
+      amplifyApi = new AmplifyData(scope, this.defaultName, {
+        apiName: this.props.name,
+        definition: amplifyGraphqlDefinition,
+        authorizationModes,
+        outputStorageStrategy: this.outputStorageStrategy,
+        functionNameMap,
+        translationBehavior: {
+          sandboxModeEnabled,
+          /**
+           * The destructive updates should be always allowed in backend definition and not to be controlled on the IaC
+           * The CI/CD check should take the responsibility to validate if any tables are being replaced and determine whether to execute the changeset
+           */
+          allowDestructiveGraphqlSchemaUpdates: true,
+        },
+      });
+    } catch (error) {
+      throw new AmplifyFault(
+        'AmplifyDataConstructFault',
+        { message: 'Failed to instantiate data construct' },
+        error as Error
+      );
+    }
+
+    /**;
      * Enable the table replacement upon GSI update
      * This is allowed in sandbox mode ONLY
      */