+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/func-src/response_generator.ts
@@ -1,5 +1,10 @@
 import { Amplify } from 'aws-amplify';
 import { downloadData, uploadData } from 'aws-amplify/storage';
+// This is needed to test dependencies that require .node files
+const {
+  utils: { generateKeyPairSync },
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+} = require('ssh2');
 /**
  * This import is for tests to use the generated type generation file.
  * Currently we only use defaultNodeFunction because node16Function has the same environment variables at runtime.
@@ -44,6 +49,7 @@ export const getResponse = async () => {
     s3TestContent: await s3RoundTrip(),
     testSecret: env.TEST_SECRET,
     testSharedSecret: env.TEST_SHARED_SECRET,
+    testNodeAddon: `${generateKeyPairSync('ed25519')}`,
   };
 };